article functionality working, ready for PR
