+++ src/scripts/articles/Article.js
@@ -3,7 +3,7 @@
 export const ArticleHTMLConverter = (artObj) => {
     return `
         <article class="article--container" id="article--${artObj.id}">
-            <h2 class="article--title">${artObj.title}</h2>
+            <h3 class="article--title">${artObj.title}</h3>
             <p class="article--timestamp">Posted: ${new Date(artObj.date).toLocaleDateString('en-US')}</p>
 
             <p class="article--synopsis">${artObj.synopsis}</p>
